power: qpnp-smbcharger: lower AICL deglitch timer during taper charge
During the transition from constant current charging to taper charging,
there is a lot of oscillation whenever the input power supply is current
limited, and the charger buck is operating in 100% duty cycle. This
causes oscillations on the USB input voltage, which in turn causes the
Fuel Gauge VADC to read faulty readings on VBAT and IBAT.

This oscillation is caused by the battery overvoltage triggering on and
off rapidly, which causes drops on USBIN. The voltage collapses are fast
enough that the AICL deglitch filters throw them out, so AICL never
lowers the input current limit.

Fix this issue by lowering the deglitch timer whenever we hit taper
charge. Also, turn on AICL reruns to allow the current to go back up to
the maximum supported after the device is out of the problem region on
the charging curve.

CRs-Fixed: 755084
Change-Id: Ie7790c40ea16a7a68cff555a8f4bcc39baf4f309
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>