power: qpnp-smbcharger: do not override APSD results
PMI8994 2.0 chargers had a hardware bug where the APSD results for a SDP
USB (Standard Downstream Port), would be 1.5A. The APSD override bit was
set for this reason. Overriding the APSD results was later found to have
a side effect of disabling AICL reruns when switching from 5V to 9V when
using a HVDCP.

In the latest PMI8994 hardware revision (PMI8994 2.0.4) this bug was
fixed. Remove the APSD override bit in order to allow AICL to rerun at
9V for HVDCPs.

CRs-Fixed: 771405
Change-Id: I6e93a0343553b9f1251554ce8b9e2cee0b46e2ad
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>