power: qpnp-smbcharger: use correct bit definitions for uv and ov
Currently, smbcharger driver checks the register for values (0x0, 0x01)
instead of the bit positions corresponding to them for uv (undervolt)
and ov(overvolt) conditions.

The effect of this error is benign until now since the code ended up
checking 0x1 = BIT(0) which is the uv case. Almost all the time
when a charger is not present the uv condition is set. Rarely does
a charger goes ov.

But nevertheless, fix the issue.

Change-Id: I8a4117fb9d483e9937a47df624c8792e03f44f1f
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>