power: qpnp-smbcharger: detect and reject faulty charger
There could be faulty chargers which can lead to frequent AICL
reruns and that can cause fuel gauge ADC to go into bad state.
To avoid that, disable the AICL rerun once a certain limit of AICL
done interrupts are received when the AICL deglitch timer is
configured for short interval. AICL rerun will be enabled back
once the AICL deglitch timer workaround is called.

If the AICL rerun is enabled, charger buck will never enter
suspend when a bad charger is detected by the hardware. Detect
these conditions in software and report it to USB power supply.

CRs-Fixed: 800230
Change-Id: Id79a99b24044e4312b515c8dba652c54cfb47346
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>