power: qpnp-smbcharger: add low-volt-dcin device tree property
Currently, we do AICL deglitch configuration dynamically for all
type of chargers whereas it should be applied specifically. As a
part of that, we need to differentiate low volt DCIN chargers.

Add a device tree property to support that requirement. If this
property is set, then AICL deglitch configuration will be applied
dynamically.

CRs-Fixed: 800438
Change-Id: Id4843b0b758652a4a99e212f286ed09118503a8e
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>