power: qpnp-smbcharger: handle the removal of OCP
Currently, we handle the removal of DCP/HVDCP in UV interrupt
handler and the removal of SDP/CDP in source detect interrupt
handler. For the cases of OCP (other charging port), we are
not handling the removal even though it might have got detected
in the source detection process. Add support to handle it.

CRs-Fixed: 828435
Change-Id: Iecccbc08fa6b8d4ab8e3963953a7b754dacbfb51
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>