power: qpnp-smbcharger: clear weak charger flag in src_det irq handler
We often see that a newly inserted charger is declared weak charger
because that status is incorrectly carried forward from the previous
charger insertion. This is especially true when usbin_uv interrupt
is triggered before or simultaneously along with src_detect interrupt.
The usbin_uv interrupt sees that the charger is present and it ends
up marking the charger as weak.

Fix this by ensuring the weak charger status is cleared upon a rising
src_detect interrupt too. The driver currently clears it only for the
falling edge case.

CRs-Fixed: 876999
Change-Id: Id370c956c74c3b972216c77718d5978f1dbd213f
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>